refactor : R 제거 #2
