+++ app/src/main/java/org/seonhwan/android/veloginmobile/ui/mypage/MypageFragment.kt
@@ -1,23 +1,22 @@
-package org.seonhwan.android.veloginmobile.presentation.Notification
+package org.seonhwan.android.veloginmobile.ui.mypage
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.seonhwan.android.veloginmobile.R
-import org.seonhwan.android.veloginmobile.databinding.FragmentNotificationBinding
+import org.seonhwan.android.veloginmobile.databinding.FragmentMypageBinding
 
-class NotificationFragment : Fragment() {
-    private var _binding: FragmentNotificationBinding? = null
-    private val binding: FragmentNotificationBinding
+class MypageFragment : Fragment() {
+    private var _binding: FragmentMypageBinding? = null
+    private val binding: FragmentMypageBinding
         get() = requireNotNull(_binding) { "_binding is not null" }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentNotificationBinding.inflate(inflater, container, false)
+        _binding = FragmentMypageBinding.inflate(inflater, container, false)
         return binding.root
     }
